Move KeybindingsRegistry from GUI to controllers layer (#589)
The keybindings module has zero Qt imports and contains file I/O
persistence logic, making it a controller-layer concern. Move it
entirely to app/controllers/keybindings.py and update all imports.
